POC/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
azure-storage-blob==12.19.0
prometheus-client==0.20.0

# בדיקות
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.0
